backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg[binary]>=3.1
python-dotenv>=1.0
dj-database-url>=2.1
gunicorn>=21.0
